Refactor: Department Service & Add Logger

+++ department-service/src/main/java/com/hrm/eureka/department/service/DepartmentService.java
@@ -5,13 +5,17 @@ import com.hrm.eureka.department.mapper.DepartmentMapper;
 import com.hrm.eureka.department.model.Department;
 import com.hrm.eureka.department.repository.DepartmentRepository;
 import jakarta.persistence.EntityNotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 public class DepartmentService {
+    private static final Logger log = LoggerFactory.getLogger(DepartmentService.class);
     private final DepartmentRepository departmentRepository;
 
     public DepartmentService(DepartmentRepository departmentRepository) {
@@ -19,19 +23,25 @@ public class DepartmentService {
     }
 
     public List<DepartmentDto> getAllDepartments(){
+        log.info("[Department Service] Getting all departments");
         return departmentRepository.findAll().stream()
                 .map(DepartmentMapper::mapToDepartmentDto)
                 .collect(Collectors.toList());
     }
 
     public DepartmentDto getDepartmentById(Long departmentId){
-        return departmentRepository.findById(departmentId)
-                .map(DepartmentMapper::mapToDepartmentDto)
-                // Throw exception if department not found
-                .orElseThrow(() -> new EntityNotFoundException("Department not found with id: " + departmentId));
+        log.info("[Department Service] Getting department by id {}", departmentId);
+
+        Optional<Department> optionalDepartment = departmentRepository.findById(departmentId);
+        if (optionalDepartment.isEmpty()) {
+            log.error("Department with id {} not found", departmentId);
+            throw new EntityNotFoundException("Department not found with id: " + departmentId);
+        }
+        return DepartmentMapper.mapToDepartmentDto(optionalDepartment.get());
     }
 
     public DepartmentDto createDepartment(DepartmentDto departmentDto){
+        log.info("[Department Service] Creating department");
         Department department = new Department();
         department.setDepartmentName(departmentDto.getDepartmentName());
         department.setDescription(departmentDto.getDescription());
@@ -40,19 +50,26 @@ public class DepartmentService {
     }
 
     public DepartmentDto updateDepartment(Long departmentId, DepartmentDto departmentDto){
-        Department department = departmentRepository.findById(departmentId)
-                .orElseThrow(() -> new EntityNotFoundException("Department not found with id: " + departmentId));
+        log.info("[Department Service] Updating department with id {}", departmentId);
+        Optional<Department> optionalDepartment = departmentRepository.findById(departmentId);
+        if (optionalDepartment.isEmpty()) {
+            log.error("Department with id {} not found", departmentId);
+            throw new EntityNotFoundException("Department not found with id: " + departmentId);
+        }
 
+        Department department = optionalDepartment.get();
         department.setDepartmentName(departmentDto.getDepartmentName());
         department.setDescription(departmentDto.getDescription());
-//        department.setManagerId(departmentDto.getManagerId());
 
         return DepartmentMapper.mapToDepartmentDto(departmentRepository.save(department));
     }
 
     public void deleteDepartment(Long departmentId){
-        Department department = departmentRepository.findById(departmentId)
-                .orElseThrow(() -> new EntityNotFoundException("Department not found with id: " + departmentId));
+        log.info("[Department Service] Deleting department with id {}", departmentId);
+        if (!departmentRepository.existsById(departmentId)) {
+            log.error("Department with id {} not found", departmentId);
+            throw new EntityNotFoundException("Department not found with id: " + departmentId);
+        }
         departmentRepository.deleteById(departmentId);
     }
